first release

+++ DBManager/dbmanager.cpp
@@ -1,34 +1,37 @@
 #include "DBManager.h"
+#include <QtSql/QSqlDatabase>
 #include <QSqlQuery>
 #include <QSqlRecord>
 
-DBManager::DBManager(QObject *parent)
+DBManager::DBManager(const QString& dbType,QObject *parent)
     : QObject(parent)
-    ,mDB(QSqlDatabase::addDatabase("QOCI"))
+	, mDB(nullptr)
 {
-   
+	mDB = new QSqlDatabase(QSqlDatabase::addDatabase("QOCI"));
+	Q_ASSERT_X(mDB != nullptr,"DBManager", "create sqldatabase fail!");
 }
 
 DBManager::~DBManager()
 {
    close();
-   QSqlDatabase::removeDatabase(mDB.connectionName());
+   QSqlDatabase::removeDatabase(mDB->connectionName());
+   delete mDB;
 }
 
 bool DBManager::isOpened() const
 {
-    return mDB.isOpen();
+    return mDB->isOpen();
 }
 
-bool DBManager::open(const QString& user,const QString& psw,const QString& address/* = "localhost"*/,const QString& dbName/* = "orcl" */)
+bool DBManager::open(const QString& user, const QString& psw, int port, const QString& address/* = "localhost"*/, const QString& dbName/* = "orcl" */)
 {
-    if(!mDB.isOpen()){
-        mDB.setHostName(address);
-        mDB.setDatabaseName(dbName);
-        mDB.setUserName(user);
-        mDB.setPassword(psw);
-        mDB.setPort(1521);
-        return mDB.open();
+    if(!mDB->isOpen()){
+		mDB->setHostName(address);
+		mDB->setDatabaseName(dbName);
+		mDB->setUserName(user);
+		mDB->setPassword(psw);
+		mDB->setPort(port);
+		return mDB->open();
     }else{
         return true;
     }
@@ -36,20 +39,23 @@ bool DBManager::open(const QString& user,const QString& psw,const QString& addre
 
 void DBManager::close()
 {
-    if(mDB.isOpen())
+	if (mDB->isOpen())
     {
-        mDB.close();
+		mDB->close();
     }
 }
 
 bool DBManager::execSql( const QString& sql )
 {
+	if (!isOpened()) return false;
     QSqlQuery query;
     return query.exec(sql);
 }
 
 bool DBManager::update( const QString& table,const QMap<QString,QVariant>& values ,const QString& _where)
 {
+	if (!isOpened()) return false;
+
     QSqlQuery query;
     QString datas;
     QList<QString> keyList = values.keys();
@@ -74,6 +80,8 @@ bool DBManager::update( const QString& table,const QMap<QString,QVariant>& value
 
 bool DBManager::remove( const QString& table,const QString& _where )
 {
+	if (!isOpened()) return false;
+
     QSqlQuery query;
     QString sql = QString("DELETE FROM %1 WHERE %2").arg(table).arg(_where);
     return query.exec(sql);
@@ -81,6 +89,8 @@ bool DBManager::remove( const QString& table,const QString& _where )
 
 bool DBManager::add( const QString& table,const QMap<QString,QVariant>& values )
 {
+	if (!isOpened()) return false;
+
     QSqlQuery query;
     QString columns,datas;
     QList<QString> keyList = values.keys();
@@ -107,6 +117,8 @@ bool DBManager::add( const QString& table,const QMap<QString,QVariant>& values )
 QList<QList<QVariant>> DBManager::query( const QString& table,QStringList& columns/*=QStringList()*/,const QString& _where )
 {
     QList<QList<QVariant>> rets;
+	if (!isOpened()) return rets;
+
     QSqlQuery query;
     QString _columns;
     if(columns.count()>0)
